requirements.txt
# Temporal-Phase Spin Retrieval System
# Requirements for Python 3.8+

# Core dependencies
numpy>=1.21.0
python-dateutil>=2.8.2

# HTTP client for LlamaStack API
httpx>=0.24.0

# FastAPI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Vector databases (optional - install as needed)
# For Chroma support:
chromadb>=0.4.0

# For PGVector support:
psycopg2-binary>=2.9.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Better CLI experience
rich>=13.0.0

